+++ shilyaev.vladislav/A1/shape.hpp
@@ -5,10 +5,10 @@
 
 class Shape {
 public:
-  virtual float getArea() const = 0;
+  virtual double getArea() const = 0;
   virtual rectangle_t getFrameRect() const = 0;
   virtual void move(const point_t& pos) = 0;
-  virtual void move(float dx, float dy) = 0;
+  virtual void move(double dx, double dy) = 0;
 };
 
 #endif
